requirements.txt
beautifulsoup4==4.12.3
python-vlc==3.0.20123
requests==2.31.0
yt-dlp==2024.3.10 